chore(runtime-html): move tests to runtime-html-browser
