+++ packages/runtime-html-browser/test/_doubles/fake-view-factory.ts
