+++ packages/runtime-html-browser/test/observation/select-value-observer.spec.ts
@@ -2,7 +2,8 @@ import { Registration } from '@aurelia/kernel';
 import { IDOM, ILifecycle, IObserverLocator, LifecycleFlags } from '@aurelia/runtime';
 import { expect } from 'chai';
 import { spy } from 'sinon';
-import { HTMLDOM, HTMLRuntimeConfiguration, SelectValueObserver } from '../../src/index';
+import { HTMLDOM, SelectValueObserver } from '../../../runtime-html/src/index';
+import { HTMLBrowserRuntimeConfiguration } from '../../src/index';
 import { _, createElement, h, verifyEqual } from '../util';
 
 type Anything = any;
@@ -10,7 +11,7 @@ type Anything = any;
 // TODO: need many more tests here, this is just preliminary
 describe('SelectValueObserver', () => {
   function createFixture(initialValue: Anything = '', options = [], multiple = false) {
-    const container = HTMLRuntimeConfiguration.createContainer();
+    const container = HTMLBrowserRuntimeConfiguration.createContainer();
     const dom = new HTMLDOM(document);
     Registration.instance(IDOM, dom).register(container, IDOM);
     const lifecycle = container.get(ILifecycle);
@@ -229,7 +230,7 @@ describe('SelectValueObserver', () => {
       type SelectValidChild = HTMLOptionElement | HTMLOptGroupElement;
 
       function createMutiSelectSut(initialValue: Anything[], options: SelectValidChild[]) {
-        const container = HTMLRuntimeConfiguration.createContainer();
+        const container = HTMLBrowserRuntimeConfiguration.createContainer();
         const dom = new HTMLDOM(document);
         Registration.instance(IDOM, dom).register(container, IDOM);
         const lifecycle = container.get(ILifecycle);
