+++ packages/runtime-html/package.json
@@ -29,12 +29,6 @@
   ],
   "scripts": {
     "lint": "cross-env tslint --project tsconfig-test.json -e \"test/**\" -t stylish",
-    "test": "cross-env TS_NODE_PROJECT=\"../../scripts/tsconfig.json\" karma start ../../scripts/karma.conf.ts --package=runtime-html --browsers=ChromeHeadlessOpt --single-run --coverage",
-    "test-firefox": "cross-env TS_NODE_PROJECT=\"../../scripts/tsconfig.json\" karma start ../../scripts/karma.conf.ts --package=runtime-html --browsers=FirefoxHeadless --single-run --coverage",
-    "test:watch": "cross-env TS_NODE_PROJECT=\"../../scripts/tsconfig.json\" karma start ../../scripts/karma.conf.ts --package=runtime-html --browsers=ChromeHeadlessOpt --coverage",
-    "test:watch:verbose": "cross-env TS_NODE_PROJECT=\"../../scripts/tsconfig.json\" karma start ../../scripts/karma.conf.ts --package=runtime-html --browsers=ChromeHeadlessOpt --coverage --reporter=mocha",
-    "test:debugger": "cross-env TS_NODE_PROJECT=\"../../scripts/tsconfig.json\" karma start ../../scripts/karma.conf.ts --package=runtime-html --browsers=ChromeDebugging",
-    "test:debugger:verbose": "cross-env TS_NODE_PROJECT=\"../../scripts/tsconfig.json\" karma start ../../scripts/karma.conf.ts --package=runtime-html --browsers=ChromeDebugging --reporter=mocha",
     "build": "tsc -b",
     "bundle": "ts-node -P ../../scripts/tsconfig.json ../../scripts/bundle.ts umd,esm,system runtime-html",
     "dev": "tsc -b -w",
@@ -48,38 +42,13 @@
     "@aurelia/runtime": "0.3.0"
   },
   "devDependencies": {
-    "@types/chai": "^4.1.7",
-    "@types/karma": "^3.0.1",
-    "@types/mocha": "^5.2.5",
     "@types/node": "^10.12.18",
-    "@types/sinon": "^7.0.2",
-    "@types/sinon-chai": "^3.2.2",
-    "chai": "^4.2.0",
     "cross-env": "^5.2.0",
-    "istanbul-instrumenter-loader": "^3.0.1",
-    "karma": "^3.1.4",
-    "karma-chai": "^0.1.0",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-coverage": "^1.1.2",
-    "karma-coverage-istanbul-reporter": "^2.0.4",
-    "karma-firefox-launcher": "^1.1.0",
-    "karma-junit-reporter": "^1.2.0",
-    "karma-mocha": "^1.3.0",
-    "karma-mocha-reporter": "^2.2.5",
-    "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^3.0.5",
-    "mocha": "^5.2.0",
-    "path": "^0.12.7",
-    "sinon": "^7.2.2",
-    "sinon-chai": "^3.3.0",
-    "ts-loader": "^5.3.2",
     "ts-node": "^7.0.1",
-    "tsconfig-paths": "^3.7.0",
     "tslib": "^1.9.3",
     "tslint": "^5.12.0",
     "tslint-microsoft-contrib": "^6.0.0",
     "tslint-sonarts": "^1.8.0",
-    "typescript": "^3.2.2",
-    "webpack": "^4.28.2"
+    "typescript": "^3.2.2"
   }
 }
