--- packages/runtime-html/tsconfig-test.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../tsconfig-test.json",
-  "compilerOptions": {
-    "emitDecoratorMetadata": false
-  },
-  "include": ["src", "test"]
-}
